+++ src/python/pants/backend/codegen/tasks/ragel_gen.py
@@ -9,19 +9,17 @@ import os
 import re
 import subprocess
 
-from twitter.common.collections import OrderedSet
-from twitter.common.dirutil import safe_mkdir_for
-
 from pants.backend.codegen.targets.java_ragel_library import JavaRagelLibrary
-from pants.backend.codegen.tasks.code_gen import CodeGen
-from pants.base.address import SyntheticAddress
+from pants.backend.codegen.tasks.simple_codegen_task import SimpleCodegenTask
+from pants.backend.jvm.targets.java_library import JavaLibrary
 from pants.base.build_environment import get_buildroot
 from pants.base.exceptions import TaskError
-from pants.binary_util import BinaryUtil
+from pants.binaries.binary_util import BinaryUtil
+from pants.util.dirutil import safe_mkdir_for
 from pants.util.memo import memoized_property
 
 
-class RagelGen(CodeGen):
+class RagelGen(SimpleCodegenTask):
   @classmethod
   def global_subsystems(cls):
     return super(RagelGen, cls).global_subsystems() + (BinaryUtil.Factory,)
@@ -51,77 +49,29 @@ class RagelGen(CodeGen):
                                      'ragel')
 
   @property
-  def javadeps(self):
-    return OrderedSet()
+  def synthetic_target_type(self):
+    return JavaLibrary
 
   def is_gentarget(self, target):
     return isinstance(target, JavaRagelLibrary)
 
-  def is_forced(self, lang):
-    return lang == 'java'
+  def execute_codegen(self, invalid_targets):
+    for target in invalid_targets:
+      output_dir = self.codegen_workdir(target)
+      for source in target.sources_relative_to_buildroot():
+        abs_source = os.path.join(get_buildroot(), source)
 
-  def genlangs(self):
-    return dict(java=lambda t: t.is_jvm)
+        output_file = os.path.join(output_dir, calculate_genfile(abs_source))
+        safe_mkdir_for(output_file)
 
-  def genlang(self, lang, targets):
-    if lang != 'java':
-      raise TaskError('Unrecognized ragel gen lang: {lang}'.format(lang=lang))
-    sources = self._calculate_sources(targets)
+        args = [self.ragel_binary, '-J', '-o', output_file, abs_source]
 
-    output_dir = self._java_out
-    lang_flag = '-J'
-
-    for source in sources:
-      output_file = os.path.join(output_dir, calculate_genfile(source))
-      safe_mkdir_for(output_file)
-
-      args = [self.ragel_binary, lang_flag, '-o', output_file, source]
-
-      self.context.log.debug('Executing: {args}'.format(args=' '.join(args)))
-      process = subprocess.Popen(args)
-      result = process.wait()
-      if result != 0:
-        raise TaskError('{binary} ... exited non-zero ({result})'.format(binary=self.ragel_binary,
-                                                                         result=result))
-
-  def _calculate_sources(self, targets):
-    sources = set()
-
-    def collect_sources(target):
-      if self.is_gentarget(target):
-        for source in target.sources_relative_to_buildroot():
-          sources.add(os.path.join(get_buildroot(), source))
-
-    for target in targets:
-      target.walk(collect_sources)
-    return sources
-
-  def createtarget(self, lang, gentarget, dependees):
-    if lang == 'java':
-      return self._create_java_target(gentarget, dependees)
-    else:
-      raise TaskError('Unrecognized ragel gen lang: {lang}'.format(lang=lang))
-
-  def _create_java_target(self, target, dependees):
-    genfiles = []
-    for source in target.sources_relative_to_source_root():
-      path = os.path.join(get_buildroot(), target.target_base, source)
-      genfile = calculate_genfile(path)
-      genfiles.append(os.path.join(self._java_out, genfile))
-
-    spec_path = os.path.relpath(self._java_out, get_buildroot())
-    spec = '{spec_path}:{name}'.format(spec_path=spec_path, name=target.id)
-    address = SyntheticAddress.parse(spec=spec)
-    tgt = self.context.add_new_target(address,
-                                      JavaRagelLibrary,
-                                      derived_from=target,
-                                      sources=genfiles,
-                                      provides=target.provides,
-                                      dependencies=self.javadeps,
-                                      excludes=target.payload.excludes)
-    for dependee in dependees:
-      dependee.inject_dependency(tgt.address)
-    return tgt
+        self.context.log.debug('Executing: {args}'.format(args=' '.join(args)))
+        process = subprocess.Popen(args)
+        result = process.wait()
+        if result != 0:
+          raise TaskError('{binary} ... exited non-zero ({result})'
+                          .format(binary=self.ragel_binary, result=result))
 
 
 def calculate_class_and_package(path):
@@ -148,7 +98,7 @@ def calculate_class_and_package(path):
 
 
 def get_filename(package, classname):
-  return "{package}/{cls}.java".format (package=package.replace(".", os.path.sep), cls=classname)
+  return "{package}/{cls}.java".format(package=package.replace(".", os.path.sep), cls=classname)
 
 
 def calculate_genfile(path):