+++ src/python/pants/binaries/binary_util.py
@@ -1,5 +1,5 @@
 # coding=utf-8
-# Copyright 2014 Pants project contributors (see CONTRIBUTORS.md).
+# Copyright 2015 Pants project contributors (see CONTRIBUTORS.md).
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 
 from __future__ import (absolute_import, division, generators, nested_scopes, print_function,
@@ -44,7 +44,6 @@ _PATH_BY_ID = {
 logger = logging.getLogger(__name__)
 
 
-# TODO(John Sirois): Extract this subsystem to its own file.
 class BinaryUtil(object):
   """Wraps utility methods for finding binary executables."""